+++ fastapi_mvc/commands/__init__.py
@@ -10,4 +10,3 @@ from fastapi_mvc.commands.base import Command
 from fastapi_mvc.commands.invoker import Invoker
 from fastapi_mvc.commands.run_generator import RunGenerator
 from fastapi_mvc.commands.run_shell import RunShell
-from fastapi_mvc.commands.new_project import GenerateNewProject