+++ tests/unit/cli/test_new.py
@@ -1,33 +1,34 @@
 import mock
 import pytest
-from fastapi_mvc.cli.new import new
+from fastapi_mvc.cli.new import get_new_cmd
 
 
 @mock.patch("fastapi_mvc.cli.new.Borg")
 def test_new_help(borg_mock, cli_runner):
-    result = cli_runner.invoke(new, ["--help"])
+    result = cli_runner.invoke(get_new_cmd(), ["--help"])
     assert result.exit_code == 0
     borg_mock.assert_not_called()
 
 
 @mock.patch("fastapi_mvc.cli.new.Borg")
 def test_new_invalid_options(borg_mock, cli_runner):
-    result = cli_runner.invoke(new, ["--not_exists"])
+    result = cli_runner.invoke(get_new_cmd(), ["--not_exists"])
     assert result.exit_code == 2
     borg_mock.assert_not_called()
 
 
-@mock.patch("fastapi_mvc.cli.new.RunShell")
-@mock.patch("fastapi_mvc.cli.new.GenerateNewProject")
+@mock.patch("fastapi_mvc.cli.new.RunGenerator")
 @mock.patch("fastapi_mvc.cli.new.Borg")
-def test_new_default_values(borg_mock, gen_mock, shell_mock, cli_runner):
-    result = cli_runner.invoke(new, ["test-project"])
-    assert result.exit_code == 0
+def test_new_default_values(borg_mock, run_mock, cli_runner, mock_project_gen):
+    with mock.patch("fastapi_mvc.cli.new.ProjectGenerator", new=mock_project_gen):
+        result = cli_runner.invoke(get_new_cmd(), ["test-project"])
+        assert result.exit_code == 0
 
     borg_mock.assert_called_once()
-    gen_mock.assert_called_once_with(
-        app_path="test-project",
+    run_mock.assert_called_once_with(
+        generator=mock_project_gen.return_value,
         options={
+            "app_path": "test-project",
             "skip_redis": False,
             "skip_aiohttp": False,
             "skip_vagrantfile": False,
@@ -36,20 +37,11 @@ def test_new_default_values(borg_mock, gen_mock, shell_mock, cli_runner):
             "skip_codecov": False,
             "skip_install": False,
             "license": "MIT",
-            "repo_url": "https://your.repo.url.here"
-        }
-    )
-    shell_mock.assert_called_once_with(
-        cmd=["make", "install"],
-        cwd="test-project",
+            "repo_url": "https://your.repo.url.here",
+        },
     )
 
-    calls = [
-        mock.call(gen_mock.return_value),
-        mock.call(shell_mock.return_value),
-    ]
-
-    borg_mock.return_value.enqueue.assert_has_calls(calls)
+    borg_mock.return_value.enqueue.assert_called_once_with(run_mock.return_value)
     borg_mock.return_value.execute.assert_called_once()
 
 
@@ -69,22 +61,20 @@ def test_new_default_values(borg_mock, gen_mock, shell_mock, cli_runner):
                 "LGPLv3+",
                 "--repo-url",
                 "https://github.com/gandalf/gondorapi",
-                "test-project"
+                "test-project",
             ],
             {
-                "options": {
-                    "skip_redis": True,
-                    "skip_aiohttp": True,
-                    "skip_vagrantfile": True,
-                    "skip_helm": True,
-                    "skip_actions": True,
-                    "skip_codecov": True,
-                    "skip_install": True,
-                    "license": "LGPLv3+",
-                    "repo_url": "https://github.com/gandalf/gondorapi"
-                },
                 "app_path": "test-project",
-            }
+                "skip_redis": True,
+                "skip_aiohttp": True,
+                "skip_vagrantfile": True,
+                "skip_helm": True,
+                "skip_actions": True,
+                "skip_codecov": True,
+                "skip_install": True,
+                "license": "LGPLv3+",
+                "repo_url": "https://github.com/gandalf/gondorapi",
+            },
         ),
         (
             [
@@ -102,44 +92,34 @@ def test_new_default_values(borg_mock, gen_mock, shell_mock, cli_runner):
                 "/home/gandalf/repos/you-shall-not-pass",
             ],
             {
-                "options": {
-                    "skip_redis": True,
-                    "skip_aiohttp": True,
-                    "skip_vagrantfile": True,
-                    "skip_helm": True,
-                    "skip_actions": True,
-                    "skip_codecov": True,
-                    "skip_install": True,
-                    "license": "LGPLv3+",
-                    "repo_url": "https://github.com/gandalf/gondorapi"
-                },
                 "app_path": "/home/gandalf/repos/you-shall-not-pass",
-            }
-        )
-    ]
+                "skip_redis": True,
+                "skip_aiohttp": True,
+                "skip_vagrantfile": True,
+                "skip_helm": True,
+                "skip_actions": True,
+                "skip_codecov": True,
+                "skip_install": True,
+                "license": "LGPLv3+",
+                "repo_url": "https://github.com/gandalf/gondorapi",
+            },
+        ),
+    ],
 )
-@mock.patch("fastapi_mvc.cli.new.RunShell")
-@mock.patch("fastapi_mvc.cli.new.GenerateNewProject")
+@mock.patch("fastapi_mvc.cli.new.RunGenerator")
 @mock.patch("fastapi_mvc.cli.new.Borg")
-def test_new_with_options(borg_mock, gen_mock, shell_mock, cli_runner, args, expected):
-    result = cli_runner.invoke(new, args)
-    assert result.exit_code == 0
+def test_new_with_options(
+    borg_mock, run_mock, cli_runner, mock_project_gen, args, expected
+):
+    with mock.patch("fastapi_mvc.cli.new.ProjectGenerator", new=mock_project_gen):
+        result = cli_runner.invoke(get_new_cmd(), args)
+        assert result.exit_code == 0
 
     borg_mock.assert_called_once()
-    gen_mock.assert_called_once_with(**expected)
+    run_mock.assert_called_once_with(
+        generator=mock_project_gen.return_value,
+        options=expected,
+    )
 
-    calls = [
-        mock.call(gen_mock.return_value),
-    ]
-
-    if "--skip-install" in args or "-I" in args:
-        shell_mock.assert_not_called()
-    else:
-        shell_mock.assert_called_once_with(
-            cmd=["make", "install"],
-            cwd="test-project",
-        )
-        calls.append(mock.call(shell_mock.return_value))
-
-    borg_mock.return_value.enqueue.assert_has_calls(calls)
+    borg_mock.return_value.enqueue.assert_called_once_with(run_mock.return_value)
     borg_mock.return_value.execute.assert_called_once()